+++ lib/impl/layer_impl.dart
@@ -1,7 +1,6 @@
 part of smartcanvas;
 
 abstract class LayerImpl extends NodeImpl implements Container<NodeImpl> {
-
   DOM.Element _element;
   LayerImpl(Layer shell): super(shell) {}