+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+#0.1.4
+## Fix missmatch between stage.children and stage.element.nodes
+## Implement svg layer suspend/resume
+
 #0.1.3
 ## If path isn't in a stage yet, use its config to calculate BBox.